Refactor: remove App.js and apply router path
