+++ src/components/layout/footer/footer.css
@@ -1,4 +1,4 @@
-@import url('../../styles/global.css');
+@import url('../../../styles/global.css');
 
 .footer {
   display: flex;
