+++ src/components/layout/nav/Nav.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
-import logo from '../../assets/common/logo.svg';
+import logo from '../../../assets/common/logo.svg';
 import './nav.css';
-import { getSampleUser } from '../../api/user';
+import { getSampleUser } from '../../../api/user';
 
 export default function Nav() {
   const [userProfile, setUserProfile] = useState({});
