+++ src/components/layout/nav/nav.css
@@ -1,4 +1,4 @@
-@import url('../../styles/global.css');
+@import url('../../../styles/global.css');
 
 .gnb {
   display: flex;
