+++ src/index.js
@@ -1,10 +1,26 @@
 import ReactDOM from 'react-dom/client';
+import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import React from 'react';
-import App from './App';
+import SharedPage from './pages/share/SharedPage';
+import LandingPage from './pages/landing/LandingPage';
+import Layout from './components/layout/Layout';
+
+const router = createBrowserRouter([
+  {
+    path: '/',
+    element: <Layout />,
+    children: [
+      {
+        path: '/',
+        element: <LandingPage />,
+      },
+      {
+        path: '/shared',
+        element: <SharedPage />,
+      },
+    ],
+  },
+]);
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <div>
-    <App />
-  </div>,
-);
+root.render(<RouterProvider router={router} />);
