+++ src/pages/landing/LandingPage.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function LandingPage() {
+  return <h1>This is Landing Page!</h1>;
+}
